scraping/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile scraping/requirements.in
#
colorama==0.4.6
    # via tqdm
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r scraping/requirements.in
smmap==5.0.2
    # via gitdb
tqdm==4.67.1
    # via -r scraping/requirements.in
tree-sitter==0.24.0
    # via
    #   -r scraping/requirements.in
    #   tree-sitter-language-pack
tree-sitter-c-sharp==0.23.1
    # via tree-sitter-language-pack
tree-sitter-embedded-template==0.23.2
    # via tree-sitter-language-pack
tree-sitter-language-pack==0.7.3
    # via -r scraping/requirements.in
tree-sitter-yaml==0.7.1
    # via tree-sitter-language-pack
